AB#32311 move AI input data provider to host app
